requirements.txt
tushare==1.4.5
akshare==1.18.4
baostock==0.8.8
pandas==2.1.4
numpy==1.26.2
pymysql==1.1.0
scikit-learn==1.3.2
xgboost==2.1.4
talib-binary==0.4.28
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
schedule==1.2.0
wxpy==0.3.9.8
plyer==2.1.0
requests==2.31.0
python-dotenv==1.0.0
